COM-1752 using mixin
